chore: update maintainer and repository metadata
Update project metadata, URLs, and maintainer information to point to the new repository. Refactor installation scripts and build configurations to improve compatibility.

+++ PKGBUILD
@@ -1,10 +1,10 @@
-# Maintainer: z-30 Working Group <dev@z30mode.org>
+# Maintainer: Paulo Mantas <paulomantas2009@gmail.com>
 pkgname=z30-transceiver
 pkgver=1.0.0
 pkgrel=1
 pkgdesc="16-MFSK Weak-Signal Digital Mode Transceiver, LDPC-SIC Decoder, CAT Controller, and DSP Suite"
 arch=('x86_64' 'aarch64' 'armv7h')
-url="https://github.com/z30mode/z30-transceiver"
+url="https://github.com/themantas1994/z-30"
 license=('MIT')
 depends=(
     'python>=3.9'
@@ -13,6 +13,7 @@ depends=(
     'python-sounddevice'
     'python-pyserial'
     'python-cffi'
+    'python-requests'
     'portaudio'
     'hamlib'
     'tk'
@@ -22,20 +23,37 @@ optdepends=(
     'npm: for building web interface'
     'python-pyaudio: alternative audio backend'
 )
-makedepends=('python-setuptools' 'python-build' 'python-installer' 'python-wheel')
-source=("$pkgname-$pkgver.tar.gz::https://github.com/z30mode/z30-transceiver/archive/v$pkgver.tar.gz")
+makedepends=('python-setuptools' 'python-build' 'python-installer' 'python-wheel' 'git')
+source=("z-30::git+https://github.com/themantas1994/z-30.git#branch=main")
 sha256sums=('SKIP')
 
 build() {
-    cd "$srcdir"
+    if [ -d "$srcdir/z-30" ]; then
+        cd "$srcdir/z-30"
+    elif [ -f "$startdir/setup.py" ]; then
+        cd "$startdir"
+    else
+        cd "$srcdir"
+    fi
     python -m build --wheel --no-isolation
 }
 
 package() {
-    cd "$srcdir"
+    if [ -d "$srcdir/z-30" ]; then
+        cd "$srcdir/z-30"
+    elif [ -f "$startdir/setup.py" ]; then
+        cd "$startdir"
+    else
+        cd "$srcdir"
+    fi
+
     python -m installer --destdir="$pkgdir" dist/*.whl
-    
+
     # Desktop integration
-    install -Dm644 z30.desktop "$pkgdir/usr/share/applications/z30.desktop"
-    install -Dm644 icon-512.svg "$pkgdir/usr/share/icons/hicolor/scalable/apps/z30.svg"
+    if [ -f z30.desktop ]; then
+        install -Dm644 z30.desktop "$pkgdir/usr/share/applications/z30.desktop"
+    fi
+    if [ -f icon-512.svg ]; then
+        install -Dm644 icon-512.svg "$pkgdir/usr/share/icons/hicolor/scalable/apps/z30.svg"
+    fi
 }
